+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project attempts to match the major and minor versions of [stactools](https://github.com/stac-utils/stactools) and increments the patch number as needed.
 
+## Unreleased
+
+### Changed
+
+- STAC item centroid uses Python `float` instead of `numpy.float64`.
+
 ## [v0.5.0] - 2024-01-17
 
 ### Added